+++ js-api/src/api/grok_api.g.ts
@@ -1453,6 +1453,7 @@ export interface IDartApi {
   grok_Get_StackTrace_Hash(stackTrace: String): any;
   grok_TreeViewGroup_Load_Sources(node: any, source: any): Promise<any>;
   grok_ReportsRule_Add_Dialog(): Promise<any>;
+  grok_Parse_Command(command: String, safe: Bool): any;
 
   // Generated from ../grok_shared/lib/grok_shared.api.g.dart
   grok_DataSourceType_Create(): any;
@@ -1558,6 +1559,7 @@ export interface IDartApi {
 
   // Generated from ../ddt/lib/ddt.api.g.dart
   grok_Tags_Create(): any;
+  grok_TagValues_Create(): any;
   grok_FuncOptions_Create(): any;
   grok_FuncParamOptions_Create(): any;
 }