+++ packages/Bio/webpack.config.js
@@ -42,9 +42,6 @@ module.exports = {
     'cash-dom': '$',
     'dayjs': 'dayjs',
     'wu': 'wu',
-    'scil': 'scil',
-    'org': 'org',
-    // 'JSDraw2': 'JSDraw2',
   },
   output: {
     filename: '[name].js',